% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwr_check_query.R
\name{bwr_check_query}
\alias{bwr_check_query}
\title{Validate a Brandwatch query}
\usage{
bwr_check_query(token = Sys.getenv("BW_TOKEN"), query = list())
}
\arguments{
\item{token}{The authentication token, acquired using bwr_auth()}

\item{query}{A list containing named parameters: query and language.}
}
\value{
Returns a list containing any errors or issues, if found
}
\description{
Validate a Brandwatch query
}
\examples{
any_issues <- bwr_check_query(query = list(query = "at_mentions:huey", language = "en"))
}
