% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{bwr_filters_get}
\alias{bwr_filters_get}
\title{Get a data frame of available filters for the Brandwatch API}
\usage{
bwr_filters_get(token = Sys.getenv("BW_TOKEN"))
}
\arguments{
\item{token}{The auth token for this user, which should be generated into an environment variable using bwr_auth()}
}
\value{
Returns a dataframe of available filters
}
\description{
See the Brandwatch documentation for more information at https://developers.brandwatch.com/docs/available-filters
}
\examples{
my_queries <- bwr_query_get(project_id = 12334534)
}
