% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_get_queries.R
\name{bwr_get_queries}
\alias{bwr_get_queries}
\title{Get a data frame of available queries for a given project}
\usage{
bwr_get_queries(token = Sys.getenv("BW_TOKEN"), project_id = NULL,
  type = NULL)
}
\arguments{
\item{token}{The auth token for this user, which should be generated into an environment variable using bwr_auth()}

\item{project_id}{The project id you'd like to return all available queries for. Obtain a data frame of projects using bwr_get_projects()}

\item{type}{(Optional) The type of query you'd like to return. If not specified, the API will return all available queries for the given project.
Currently, results don't seem to be affected by the type parameter - the API appears to return full results in all cases.}
}
\value{
Returns a dataframe of queries matching the project.
}
\description{
See the Brandwatch documentation for more information at https://developers.brandwatch.com/docs/retrieving-queries.
}
\examples{
my_queries <- bwr_get_queries(project_id = 12334534)
}
