% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mentions.R
\name{bwr_mentions_get}
\alias{bwr_mentions_get}
\title{Get a list of mentions for the specified query and project}
\usage{
bwr_mentions_get(token = Sys.getenv("BW_TOKEN"), project_id = NULL,
  query_id = NULL, filters = NULL, date_range = c(Sys.Date() - 31,
  Sys.Date() - 1), page = NULL)
}
\arguments{
\item{token}{The authentication token, acquired using bwr_auth()}

\item{project_id}{The project id in which the specified query is contained. Obtain a list of project IDs using bwr_get_projects().}

\item{query_id}{The id of the query you'd like to run. Obtain a list of queries for any specified project using bwr_get_queries().}

\item{filters}{(Optional) A list of key-value pairs to filter the mentions query by. Use the bwr_filters_get() function to find out all available filters.}

\item{date_range}{A character vector containing 2 date values in YYYY-mm-dd format. The first value is the beginning of your desired date range and the second value is the end of the date range.}

\item{page}{(Optional) The page number for which to return results. This is needed only when R needs to iterate through multiple pages of results from the Brandwatch API. It is recommended that you keep this value set as zero.}
}
\value{
Returns a data frame containing all results
}
\description{
Get a list of mentions for the specified query and project
}
\examples{
my_project <- bwr_get_projects()$id[1]
my_query <- bwr_get_queries(project_id = my_project)$id[1]
my_mentions <- bwr_mentions_get(
project_id = my_project,
query_id = my_query,
date_range = c("2018-01-01", "2018-02-01"))

}
