% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{bwr_query_check}
\alias{bwr_query_check}
\title{Validate a Brandwatch query}
\usage{
bwr_query_check(query = list(), token = Sys.getenv("BW_TOKEN"))
}
\arguments{
\item{query}{A list containing named parameters: query and language.}

\item{token}{The authentication token, acquired using bwr_auth()}
}
\value{
Returns a list containing any errors or issues, if found
}
\description{
Validate a Brandwatch query
}
\examples{
\dontrun{any_issues <- bwr_query_check(query = list(query = 'at_mentions:huey', language = 'en'))}
}
