% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{bwr_rule_get}
\alias{bwr_rule_get}
\title{Get a data frame of available rules for a given project}
\usage{
bwr_rule_get(project_id = NULL, token = Sys.getenv("BW_TOKEN"))
}
\arguments{
\item{project_id}{The project id you'd like to return all available rules for. Obtain a data frame of projects using bwr_get_projects()}

\item{token}{The auth token for this user, which should be generated into an environment variable using bwr_auth()}
}
\value{
Returns a dataframe of rules matching the project. Due to the structure of the API response, some of the dataframe columns are list-columns (i.e. nested dataframes).
}
\description{
See the Brandwatch documentation for more information at https://developers.brandwatch.com/docs/retrieving-rules.
}
\examples{
\dontrun{my_rules <- bwr_rules_get(project_id = 12334534)}
}
