% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{bwr_tag_create}
\alias{bwr_tag_create}
\title{Upload a new tag to Brandwatch}
\usage{
bwr_tag_create(token = Sys.getenv("BW_TOKEN"), project_id = NULL,
  name = NULL)
}
\arguments{
\item{token}{The authentication token, obtained using bwr_auth()}

\item{project_id}{The project ID into which you'd like to insert a tag}

\item{name}{The name of the tag}
}
\value{
Returns a list containing the JSON response from the server.
}
\description{
Refer to https://developers.brandwatch.com/docs/creating-tags for more information. Unless otherwise specified, provide a single string argument.
}
\examples{
\dontrun{bwr_tag_create(project_id = 12423432, name = 'Quick test tag')}
}
