% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{bwr_tag_get}
\alias{bwr_tag_get}
\title{Get a data frame of available tags for a given project}
\usage{
bwr_tag_get(token = Sys.getenv("BW_TOKEN"), project_id = NULL)
}
\arguments{
\item{token}{The auth token for this user, which should be generated into an environment variable using bwr_auth()}

\item{project_id}{The project id you'd like to return all available tags for. Obtain a data frame of projects using bwr_get_projects()}
}
\value{
Returns a dataframe of tags matching the project.
}
\description{
See the Brandwatch documentation for more information at https://developers.brandwatch.com/docs/retrieving-tags.
}
\examples{
my_tags <- bwr_tag_get(project_id = 12334534)
}
