% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querygroups.R
\name{bwr_querygrp_get}
\alias{bwr_querygrp_get}
\title{Get a data frame of Brandwatch query groups in the specified project}
\usage{
bwr_querygrp_get(project_id = NULL, token = Sys.getenv("BW_TOKEN"))
}
\arguments{
\item{project_id}{The project ID whose query groups you'd like to return}

\item{token}{The authentication token, obtained using bwr_auth()}
}
\value{
Returns a data frame of the query groups for the given project, if any exist.
}
\description{
Get a data frame of Brandwatch query groups in the specified project
}
\examples{
\dontrun{groups <- bwr_querygrp_get(project_id = 21343242)}
}
