% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{bwr_wf_get}
\alias{bwr_wf_get}
\title{Get a list of workflows for the specified query and project}
\usage{
bwr_wf_get(project_id = NULL, token = Sys.getenv("BW_TOKEN"))
}
\arguments{
\item{project_id}{The project id in which the specified workflow is contained. Obtain a list of project IDs using bwr_get_projects().}

\item{token}{The authentication token, acquired using bwr_auth()}
}
\value{
Returns a list containing all results
}
\description{
Refer to https://developers.brandwatch.com/docs/retrieving-workflow for more information.
}
\examples{
\dontrun{my_project <- bwr_get_projects()$id[1]
my_workflow <- bwr_wf_get(project_id = my_project)}

}
